[IMP] mrp: compute quantities speedup
The computation of kit quantities was not efficient. A loop on each product
of the recordset to consider performed. Each time a product was seen as 'not a kit',
a new computation on the complete recordset was done. This leads to a N^2 complexity

This commit only compute stock quantities for non kit product in one batch

opw-2070768

closes odoo/odoo#37205

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>