[IMP] sale_expense: employee can see all name_get of Sales Order
As f8dfe3a52be5e80f7fa6539163d3a65bdc2e8668 wants to allow simple empoyee
to select any SO on which the expense should be reinvoice, we need a little
adjustement. Indeed, before this commit, the ir.rule of current user are
still apply (bad comprehension of `name_get_uid` parameter), so a salesman
will only be alble to reinvoice on its own SO while an employee can do the
same on any SO.
This commit fixes this strange use case with rco's advice to make this
mecanism work in a cleaner way.
To restrict this `name_search` done in `sudo`, we apply a context key, and
an hardcoded domain (combine with a AND) to minimize the divulgation.

Also, we want the saleman to be able to open and see the SO.

Task-1843241

closes odoo/odoo#35938

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>