[FIX] website_event_track: rename Partner to Speaker and remove duplicated fields
Before this patch, the word "Partner" wasn't specific enough to make the user
know that it was referring to the speaker. This was OK in the form, bug i.e.
in the calendar view it's very confusing.

This change comes from b8cc28ac567afe781515c2a5ad0c0fe48031263f and we revert back to Speaker labelling.

Also, in the form view, the partner field was duplicated, contributing even
more to the confusion. It probably comes from 8baefb08a048ef69c761c679eaee1e4fe3ba4cee that duplicated the field
probably because of a bad conflict resolution.

Now partner is labelled Speaker and multi company domain is correctly taken
into account.

closes odoo/odoo#37781

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>