[FIX] pos_restaurant: edition of table
When you were trying to edit the table map of your pos, you were getting
a traceback because, it was getting information from a table that
is not in the list of table.

So we are now correctly set the table in the list when it is created and
duplicated.

TASK-ID: 2057368