[FIX] point_of_sale: apply tax repartition lines on amount
Since implementation of repartition lines on tax in accounting, the
amount on the tax is not always the effective amount to apply on price
of a product. You can have a tax with amount of 20% for example, and two
repartition lines of 100% and -100%, that will lead to have to apply 0%
on the tax.

So we are computing the effective amount on the tax by multiplying the
amount of tax by the sum of repartition lines factor.

closes odoo/odoo#36140

Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>